+++ dio/CHANGELOG.md
@@ -5,13 +5,14 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 ## Unreleased 6.0.0
 
-- The minimum supported Dart version has been bumped from `2.15.0` to `2.17.0`.
+- The minimum supported Dart version has been bumped from `2.15.0` to `2.19.0`.
 - Remove `DefaultHttpClientAdapter` which was deprecated in `5.0.0`.
 - Remove `IOHttpClientAdapter.onHttpClientCreate` which was deprecated in `5.2.0`
 - Remove `DioError` and `DioErrorType` which was deprecated in `5.2.0`.
 - Remove `DefaultTransformer` which was deprecated in `5.0.0`.
 - `IOHttpClientAdapter` no longer sets a custom `HttpClient.idleTimeout`. A custom `HttpClient` can be provided via
   `IOHttpClientAdapter.createHttpClient` if customisation is required.
+- Make use of Isolate.run for the BackgroundTransformer
 
 ## Unreleased