+++ dio/CHANGELOG.md
@@ -5,6 +5,10 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 ## Unreleased
 
+*None.*
+
+## 5.3.4
+
 - Raise warning for `Map`s other than `Map<String, dynamic>` when encoding request data.
 - Improve exception messages.
 - Allow `ResponseDecoder` and `RequestEncoder` to be async.